+++ .eslintrc.js
@@ -1,4 +1,30 @@
 module.exports = {
-  root: true,
-  extends: '@react-native-community',
+    env: {
+        es6: true,
+    },
+    extends: ['airbnb', 'prettier', 'prettier/react'],
+    globals: {
+        Atomics: 'readonly',
+        SharedArrayBuffer: 'readonly',
+        __DEV__: 'readonly',
+    },
+    parser: 'babel-eslint',
+    parserOptions: {
+        ecmaFeatures: {
+            jsx: true,
+        },
+        ecmaVersion: 2018,
+        sourceType: 'module',
+    },
+    plugins: ['react', 'prettier'],
+    rules: {
+        'prettier/prettier': 'error',
+        'react/jsx-filename-extension': [
+            'warn',
+            {
+                extensions: ['.jsx', '.js'],
+            },
+        ],
+        'import/prefer-default-export': 'off',
+    },
 };
