+++ src/config/ReactotronConfig.js
@@ -1,0 +1,11 @@
+import Reactotron from 'reactotron-react-native';
+
+if (__DEV__) {
+    const tron = Reactotron.configure({ host: '192.168.15.5' })
+        .useReactNative()
+        .connect();
+
+    console.tron = tron;
+
+    tron.clear();
+}
